Fixed a whole lot of the logic needed in the DB

+++ src/main/java/no/stonedstonar/BookREST/model/Book.java
@@ -217,7 +217,7 @@ public class Book {
      * @param numberOfPages the numberofpages to check.
      */
     private void checkNumberOfPages(int numberOfPages){
-        checkIfLongIsAboveZero((long) numberOfPages, "number of pages");
+        checkIfLongIsAboveZero(numberOfPages, "number of pages");
     }
 
     /**
